[#11] TREE형태 데이터의 기본적인 CRUD루틴을 작성하여 올림. 해당 방식으로 개발을 강제할 것.

+++ src/main/java/com/nudo/gg/sample/service/SampleServiceImpl.java
@@ -46,6 +46,9 @@ public class SampleServiceImpl implements SampleService {
 
 	@Override
 	public int delete(Long id) {
+		if( ObjectUtil.isEmpty(id) ) {
+			throw new BizException("ERR.COM.002");
+		}
 		return sampleMapper.delete(id);
 	}
